Implemented queue pausing
A paused queue will not tick any of its children. This does not
mean that the children's next tick time is updated, so if you have
a queue that is paused for a long period of time you are then likely
to also have a lot of items that need ticking.  The max tick per call
should help you deal with that, however.