TickedQueue will not tick removed TickedQueueItems
Ran into an edge case with UnitySteer where a vehicle was being disabled
while the TickedQueue's work queue was being run through.  This had the
side effect that the item was removed from the main queue, but since it
was a "loop" item and already in the work queue, the Update method then
re-added it to the queue after ticking it.

Removing a TickedQueueItem from a queue will mark it as inactive. The
queue will now ignore inactive items.

Expanded tests.